+++ yaguroute/src/main/java/com/baseball/service/domain/Post.java
@@ -11,7 +11,8 @@ public class Post {
 	private String postContents;
 	private Timestamp postDate;
 	private int postViews;
-	private int postType; //0:잡담 1:응원 2:중고판매 3:중고구매 4:notice
+	private int postType; //0:잡담 1:응원 2:중고판매 3:중고구매
+	private int notice;  //0 or 1
 	private int postLikes;
 	private int postDislikes;
 	
@@ -74,6 +75,13 @@ public class Post {
         this.postType = postType;
     }
 
+    public int getNotice() {
+        return notice;
+    }
+    public void setNotice(int notice) {
+        this.notice = notice;
+    }
+    
     public int getPostLikes() {
         return postLikes;
     }
@@ -99,6 +107,7 @@ public class Post {
                 ", postDate=" + postDate +
                 ", postViews=" + postViews +
                 ", postType='" + postType + '\'' +
+                ", notice=" + notice +
                 ", postLikes=" + postLikes +
                 ", postDislikes=" + postDislikes +
                 ']';
